Auth version

+++ src/main/java/br/com/felipesoarestech/api/cliente/domain/dto/ProductResponseDTO.java
@@ -2,7 +2,7 @@ package br.com.felipesoarestech.api.cliente.domain.dto;
 
 import br.com.felipesoarestech.api.cliente.domain.model.Product;
 
-public record ProductResponseDTO(String id, String name, Integer price) {
+public record ProductResponseDTO(String id, String name, Double price) {
     public ProductResponseDTO(Product product){
         this(product.getId(), product.getName(), product.getPrice());
     }
